+++ src/objects/core/tests/factories.py
@@ -26,7 +26,7 @@ class ObjectTypeVersionFactory(factory.django.DjangoModelFactory[ObjectTypeVersi
         {
             "type": "object",
             "title": "Tree",
-            "$schema": "http://json-schema.org/draft-07/schema#",
+            "$schema": "https://json-schema.org/draft/2020-12/schema#",
             "required": ["diameter"],
             "properties": {
                 "diameter": {"type": "integer", "description": "size in cm."},