+++ src/objects/tests/v2/test_objecttypeversion_api.py
@@ -14,7 +14,7 @@ from .utils import reverse
 JSON_SCHEMA = {
     "type": "object",
     "title": "Tree",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "https://json-schema.org/draft/2020-12/schema#",
     "required": ["diameter"],
     "properties": {
         "diameter": {"type": "integer", "description": "size in cm."},
@@ -107,7 +107,7 @@ class ObjectTypeVersionAPITests(TokenAuthMixin, APITestCase):
         new_json_schema = {
             "type": "object",
             "title": "Tree",
-            "$schema": "http://json-schema.org/draft-07/schema#",
+            "$schema": "https://json-schema.org/draft/2020-12/schema#",
             "required": ["diameter"],
             "properties": {"diameter": {"type": "number"}},
         }